test: Move test case to DeploymentRejectionTest
