+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/deployment/DeploymentRejectionTest.java
@@ -23,6 +23,7 @@ import io.camunda.zeebe.protocol.record.RejectionType;
 import io.camunda.zeebe.protocol.record.ValueType;
 import io.camunda.zeebe.protocol.record.intent.CommandDistributionIntent;
 import io.camunda.zeebe.protocol.record.intent.DeploymentIntent;
+import io.camunda.zeebe.protocol.record.intent.MessageStartEventSubscriptionIntent;
 import io.camunda.zeebe.protocol.record.intent.ProcessIntent;
 import io.camunda.zeebe.protocol.record.value.DeploymentRecordValue;
 import io.camunda.zeebe.test.util.record.RecordingExporter;
@@ -567,6 +568,34 @@ public class DeploymentRejectionTest {
         .contains(
             "- ERROR: Names must not be longer than the configured max-name-length of %s characters"
                 .formatted(MAX_NAME_FIELD_LENGTH));
+  }
+
+  @Test
+  public void shouldRejectDeploymentIfMessageStartEventNameIsTooLong() {
+    // given
+    final var tooLongMessageName = "a".repeat(MAX_NAME_FIELD_LENGTH + 1);
+    final var process =
+        Bpmn.createExecutableProcess("processId")
+            .startEvent("startEventId")
+            .message(m -> m.name(tooLongMessageName).id("startmsgId"))
+            .endEvent()
+            .done();
+
+    // when
+    final var rejectedDeployment =
+        ENGINE.deployment().withXmlResource("process.bpmn", process).expectRejection().deploy();
+
+    // then
+    assertThat(rejectedDeployment.getRejectionReason())
+        .contains("Names must not be longer than the configured max-name-length")
+        .contains(String.valueOf(MAX_NAME_FIELD_LENGTH));
+    assertThat(
+            RecordingExporter.records()
+                .limit(r -> r.getPosition() >= rejectedDeployment.getPosition())
+                .messageStartEventSubscriptionRecords()
+                .withIntent(MessageStartEventSubscriptionIntent.CREATED)
+                .exists())
+        .isFalse();
   }
 
   @Test
